Implement Question authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -34,6 +34,9 @@ class QuestionsController extends Controller
 
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied!");   
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -45,6 +48,9 @@ class QuestionsController extends Controller
 
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied!");   
+        }
         $question->delete();
         return redirect('/questions')->with('success', 'Your question has been deleted.');
     }
